API

OnlinePortfolioSelection.OPSAlgorithm — Type

OPSAlgorithm{T<:Float64}

A OPSAlgorithm object that contains the result of running the algorithm.

Fields

n_asset::Int: Number of assets in the portfolio.

b::Matrix{T}: Weights of the created portfolios.

budgets::Vector{T}: Budget during the investment horizon.

alg::String: Name of the algorithm.

The formula for calculating the cumulative return of the portfolio is as follows:

\[{S_n} = {S_0}\prod\limits_{t = 1}^T {\left\langle {{b_t},{x_t}} \right\rangle }\]

where $S₀$ is the initial budget, $n$ is the investment horizon, $b_t$ is the vector of weights of the period $t$, and $x_t$ is the relative price of the $t$-th period.

OnlinePortfolioSelection.OPSMetrics — Type

OPSMetrics(Sn::Vector{T}, APY::T, Ann_Sharpe::T, MDD::T, Calmar::T) where {T<:Float64}

A struct to store the metrics of the OPS algorithm.

Fields

Sn::Vector{T}: the cumulative return of investment during the investment period.

APY::T: the Annual Percentage Yield (APY) of investment.

Ann_Std::T: the Annualized Standard Deviation (σₚ) of investment.

Ann_Sharpe::T: the Annualized Sharpe Ratio (SR) of investment.

MDD::T: the Maximum Drawdown (MDD) of investment.

Calmar::T: the Calmar Ratio of investment.

OnlinePortfolioSelection.OPSMetrics — Method

OPSMetrics(
  weights::Matrix{T},
  rel_pr::Matrix{T};
  init_inv::T=1.,
  Rf::T=0.02
  dpy::S=252
) where {T<:Float64, S<:Int}

Calculate the metrics of an OPS algorithm.

Arguments

weights::Matrix{T}: the weights of the portfolio.

rel_pr::Matrix{T}: the relative price of the stocks.

init_inv::T=1: the initial investment.

Rf::T=0.02: the risk-free rate of return.

dpy::S=252: the number of days in a year.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

::OPSMetrics: the metrics of the OPS algorithm.

OnlinePortfolioSelection.APY — Method

APY(Sn::Float64, n_periods::S; dpy::S=252) where S<:Int

Calculate the Annual Percentage Yield (APY) of investment.

Arguments

Sn::Float64: the cumulative return of investment.

n_periods::S: the number investment periods.

dpy::S=252: the number of days in a year.

Returns

::Float64: the APY of investment.

OnlinePortfolioSelection.Ann_Sharpe — Method

Ann_Sharpe(APY::T, Rf::T, sigma_prtf::T) where T<:Float64

Calculate the Annualized Sharpe Ratio of investment.

Arguments

APY::T: the APY of investment.

Rf::T: the risk-free rate of return.

sigmaprtf::T: the standard deviation of the portfolio ``\sigmap``.

Returns

::Float64: the Annualized Sharpe Ratio of investment.

OnlinePortfolioSelection.CORNK — Method

CORNK(
  adj_close::Matrix{Float64},
  horizon::T,
  k::T,
  w::T,
  p::T;
  init_budg=1
) where T<:Int

Run CORN-K algorithm.

Arguments

adj_close::Matrix{Float64}: Adjusted close prices of assets.

horizon::T: The number of periods to invest.

k::T: The number of top experts to be selected.

w::T: maximum length of time window to be examined.

p::T: maximum number of correlation coefficient thresholds.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): An object of type OPSAlgorithm.

Reference

[1] CORN: Correlation-driven nonparametric learning approach for portfolio selection

Examples

julia> using OPS

julia> adj_close = rand(5, 100);

julia> model = CORNK(adj_close, 10, 3, 5, 3);

julia> model.budgets
11-element Vector{Float64}:
    1.0
    0.9245593797606169
    1.2661989745602595
    1.9287471593384662
   13.920366533737552
   34.95772014332745
   37.275738278098444
  151.14402066138192
  186.3883800018429
  208.7067031619314
 8593.328131580158

julia> model.type
"CORN-K"

OnlinePortfolioSelection.CORNU — Method

CORNU(
  adj_close::Matrix{T},
  horizon::M,
  w::M;
  rho::T,
  init_budg=1
) where {T<:Float64, M<:Int}

Run CORN-U algorithm.

Arguments

adj_close::Matrix{T}: Adjusted close prices of assets.

horizon::M: The number of periods to invest.

w::M: maximum length of time window to be examined.

rho::T=0.2: The correlation coefficient threshold.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): An object of type OPSAlgorithm.

Reference

[1] CORN: Correlation-driven nonparametric learning approach for portfolio selection

Examples

julia> using OnlinePortfolioSelection

julia> adj_close = rand(5, 100);

julia> model = CORNU(adj_close, 10, 5, 0.5);

julia> model.budgets
11-element Vector{Float64}:
   1.0
   2.712148607094074
   3.9049548118405513
   9.079571226009149
  15.550023500498401
  60.15172309699462
  32.32354281989587
 110.76798827023994
 148.34814522333753
 206.383234246206
 463.69154027959627

julia> model.type
"CORN-U"

OnlinePortfolioSelection.CRP — Method

CRP(adj_close::Matrix{T}, init_budg=1) where T<:Float64

Run Constant Rebalanced Portfolio (CRP) algorithm.

Arguments

adj_close::Matrix{Float64}: adjusted close prices

init_budg::Float64: initial budget

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): OPSAlgorithm object

References

[1] Universal Portfolios

Example

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> crp = CRP(adj_close);

julia> crp.b
3×10 Matrix{Float64}:
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333

julia> sum(crp.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.Calmar — Method

Calmar(APY::T, MDD::T) where T<:Float64

Calculate the Calmar Ratio of investment.

Arguments

APY::T: the APY of investment.

MDD::T: the MDD of investment.

Returns

::Float64: the Calmar Ratio of investment.

OnlinePortfolioSelection.DRICORNK — Method

DRICORNK(
  adj_close::Matrix{T},
  adj_close_market::Vector{T},
  horizon::M,
  k::M,
  w::M,
  p::M;
  lambda::T=1e-3,
  init_budg=1
) where {T<:Float64, M<:Int}

Run the DRICORNK algorithm.

Arguments

adj_close::Matrix{T}: A matrix of adjusted close prices of the assets.

adj_close_market::Vector{T}: A vector of adjusted close prices of the market in the same period.

horizon::M: The investment horizon.

k::M: The number of experts.

w::M: maximum length of time window to be examined.

p::M: maximum number of correlation coefficient thresholds.

lambda::T=1e-3: The regularization parameter.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): An object of type OPSAlgorithm.

Reference

[1] DRICORN-K: A Dynamic RIsk CORrelation-driven Non-parametric Algorithm for Online Portfolio Selection

Example

julia> using OPS

julia> stocks_adj, market_adj = rand(10, 100), rand(100);

julia> dricornk = DRICORNK(stocks_adj, market_adj, 5, 2, 4, 3);

julia> dricornk.budgets
6-element Vector{Float64}:
   1.0
   3.473117165390519
   5.68035943256069
  16.98710435236131
 108.58630004976507
 323.6896116088503

OnlinePortfolioSelection.EG — Method

EG(adj_close::Matrix{Float64}, init_budg=1., eta=0.05)

Exponential Gradient (EG) algorithm.

Calculate the Exponential Gradient (EG) weights and budgets using the given historical prices and parameters and return an EG object.

Arguments

adj_close::Matrix{Float64}: Historical adjusted close prices.

init_budg=1.: Initial budget.

eta=0.05: Learning rate.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): OPSAlgorithm object.

References

[1] On-Line Portfolio Selection Using Multiplicative Updates](https://onlinelibrary.wiley.com/doi/10.1111/1467-9965.00058)

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> eg = EG(adj_close);

julia> eg.b
3×10 Matrix{Float64}:
 0.333333  0.333119  0.333296  0.333232  0.33327   0.333276  0.333201  0.333171  0.332832  0.332789
 0.333333  0.333436  0.333274  0.333485  0.333481  0.333359  0.333564  0.333477  0.333669  0.333835
 0.333333  0.333445  0.33343   0.333283  0.333249  0.333365  0.333234  0.333353  0.333499  0.333377

julia> sum(eg.b, dims=1) .|> isapprox(1.0) |> all
true

OnlinePortfolioSelection.MDD — Method

MDD(Sn::Vector{T}) where T<:Float64

Calculate the Maximum Drawdown (MDD) of investment.

Arguments

Sn::Vector{T}: the cumulative return of investment during the investment period.

Returns

::Float64: the MDD of investment.

OnlinePortfolioSelection.RPRT — Method

RPRT(
  adj_close::Matrix{T};
  w::Int64=5,
  init_budg::Int=1.,
  theta::T=0.8,
  epsilon=50
) where T<:Float64

Run RPRT algorithm.

Arguments

adj_close::Matrix{T}: Adjusted close prices of assets.

w::Int64=5: maximum length of time window to be examined.

init_budg::Int=1.: The initial budget for investment.

theta::T=0.8: The threshold for the relative price.

epsilon=50: The threshold for the condition of the portfolio.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): An object of type OPSAlgorithm.

Reference

[1] Reweighted Price Relative Tracking System for Automatic Portfolio Optimization

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> rprt = RPRT(adj_close);

julia> rprt.b
3×10 Matrix{Float64}:
 0.333333  0.333333  0.333333  0.333333  0.0       0.0369806  0.315476  0.382559  0.942964   0.678569
 0.333333  0.333333  0.333333  0.333333  0.464174  0.694045   0.137325  0.392694  0.0313677  0.0
 0.333333  0.333333  0.333333  0.333333  0.535826  0.268975   0.547199  0.224747  0.0256685  0.321431

julia> sum(rprt.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.Sn — Method

Sn(
  rel_pr::AbstractMatrix{T},
  w::AbstractMatrix{T},
  init_budg
) where T<:Float64

Calculate the cumulative return of portfolio.

Arguments

rel_pr::AbstractMatrix{T}: The relative prices of the assets in each period.

w::AbstractMatrix{T}: The weights of the assets in each period.

init_budg: The initial budget.

Beware!

rel_pr and w should be a matrix of size n_assets × n_periods.

OnlinePortfolioSelection.Sn — Method

Sn(weights::Matrix{T}, rel_pr::Matrix{T}; init_inv::T=1.) where T<:Float64

Calculate the cumulative return of investment during a period of time.

Arguments

weights::Matrix{T}: the weights of the portfolio.

rel_pr::Matrix{T}: the relative price of the stocks.

init_inv::T=1: the initial investment.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

all_sn::Vector{T}: the cumulative return of investment during the investment period.

OnlinePortfolioSelection.UP — Method

UP(
  adj_close::Matrix{Float64};
  init_budg=1.,
  eval_points::Int=10^4,
  leverage=1.,
  frequency::Int=1,
  min_history::Int=0
)

Universal Portfolio (UP) algorithm.

Calculate the Universal Portfolio (UP) weights and budgets using the given historical prices and parameters.

Arguments

adj_close::Matrix{Float64}: Historical adjusted close prices.

init_budg=1.: Initial budget.

eval_points::Int=10^4: Number of evaluation points.

leverage=1.: Leverage value.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, budgets, alg): OPSAlgorithm object.

References

[1] Universal Portfolios

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 30))

julia> up = UP(adj_close);

julia> up.b
3×30 Matrix{Float64}:
 0.333333  0.331149  0.33204   0.331716  …  0.326788  0.325788  0.325829  0.326222
 0.333333  0.336058  0.335239  0.336304     0.343405  0.342161  0.342283  0.340693
 0.333333  0.332793  0.33272   0.331981     0.329807  0.332051  0.331888  0.333086

julia> sum(up.b, dims=1) .|> isapprox(1.) |> all
true