Types

OnlinePortfolioSelection.KMDLOG — Type

KMDLOG<:ClusLogVariant

KMDLOG is a concrete type used to represent the KMDLOG Model. Also, see KMNLOG.

OnlinePortfolioSelection.KMNLOG — Type

KMNLOG<:ClusLogVariant

KMNLOG is a concrete type used to represent the KMNLOG Model. Also, see KMDLOG.

OnlinePortfolioSelection.OPSAlgorithm — Type

OPSAlgorithm{T<:AbstractFloat}

A OPSAlgorithm object that contains the result of running the algorithm.

Fields

n_asset::Int: Number of assets in the portfolio.

b::Matrix{T}: Weights of the created portfolios.

alg::String: Name of the algorithm.

OnlinePortfolioSelection.OPSMetrics — Type

OPSMetrics{T<:AbstractFloat}

A struct to store the metrics of the OPS algorithm.

Fields

Sn::Vector{T}: the cumulative return of investment during the investment period.

MER::T: the investments's Mean excess return (MER).

APY::T: the Annual Percentage Yield (APY) of investment.

Ann_Std::T: the Annualized Standard Deviation (σₚ) of investment.

Ann_Sharpe::T: the Annualized Sharpe Ratio (SR) of investment.

MDD::T: the Maximum Drawdown (MDD) of investment.

Calmar::T: the Calmar Ratio of investment.

OnlinePortfolioSelection.OPSMetrics — Method

OPSMetrics(
  weights::AbstractMatrix{T},
  rel_pr::AbstractMatrix{T};
  init_inv::T=1.,
  Rf::T=0.02
  dpy::S=252,
  v::T=0.
  dpy::S=252
) where {T<:AbstractFloat, S<:Int}

Calculate the metrics of an OPS algorithm. Also, see sn, mer, ann_std, apy, ann_sharpe, mdd, and calmar.

Arguments

weights::AbstractMatrix{T}: the weights of the portfolio.

rel_pr::AbstractMatrix{T}: the relative price of the stocks.

Keyword Arguments

init_inv::T=1: the initial investment.

Rf::T=0.02: the risk-free rate of return.

dpy::S=252: the number of days in a year.

v::T=0.: the transaction cost rate.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

::OPSMetrics: An OPSMetrics object.

OnlinePortfolioSelection.PAMR — Type

PAMR<: PAMRModel

Create a PAMR object. Also, see PAMR1, and PAMR2.

Example

model = PAMR()

OnlinePortfolioSelection.PAMR1 — Type

PAMR1{T<:AbstractFloat}<: PAMRModel

Create a PAMR1 object. Also, see PAMR, and PAMR2.

Keyword Arguments

C::AbstractFloat=1.: Aggressiveness parameter.

Example

model = PAMR1(C=0.02)

OnlinePortfolioSelection.PAMR2 — Type

PAMR2{T<:AbstractFloat}<: PAMRModel

Create a PAMR2 object. Also, see PAMR, and PAMR1.

Keyword Arguments

C::AbstractFloat=1.: Aggressiveness parameter.

Example

model = PAMR2(C=0.02)