API

OnlinePortfolioSelection.ann_sharpe — Method

ann_sharpe(APY::T, Rf::T, sigma_prtf::T) where T<:Float64

Calculate the Annualized Sharpe Ratio of investment.

Arguments

APY::T: the APY of investment.

Rf::T: the risk-free rate of return.

sigmaprtf::T: the standard deviation of the portfolio ``\sigmap``.

Returns

::Float64: the Annualized Sharpe Ratio of investment.

OnlinePortfolioSelection.ann_std — Method

ann_std(cum_ret::Vector{Float64}; dpy)

Calculate the Annualized Standard Deviation (σₚ) of portfolio.

Arguments

cum_ret::Vector{Float64}: the cumulative return of investment during the investment period.

dpy: the number of days in a year.

Returns

::Float64: the Annualized Standard Deviation (σₚ) of portfolio.

OnlinePortfolioSelection.apy — Method

apy(Sn::Float64, n_periods::S; dpy::S=252) where S<:Int

Calculate the Annual Percentage Yield (APY) of investment.

Arguments

Sn::Float64: the cumulative return of investment.

n_periods::S: the number investment periods.

dpy::S=252: the number of days in a year.

Returns

::Float64: the APY of investment.

OnlinePortfolioSelection.calmar — Method

calmar(APY::T, MDD::T) where T<:Float64

Calculate the Calmar Ratio of investment.

Arguments

APY::T: the APY of investment.

MDD::T: the MDD of investment.

Returns

::Float64: the Calmar Ratio of investment.

OnlinePortfolioSelection.cornk — Method

cornk(
  adj_close::Matrix{Float64},
  horizon::T,
  k::T,
  w::T,
  p::T;
  init_budg=1
) where T<:Int

Run CORN-K algorithm.

Arguments

adj_close::Matrix{Float64}: Adjusted close prices of assets.

horizon::T: The number of periods to invest.

k::T: The number of top experts to be selected.

w::T: maximum length of time window to be examined.

p::T: maximum number of correlation coefficient thresholds.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Reference

[1] CORN: Correlation-driven nonparametric learning approach for portfolio selection

Examples

julia> using OPS

julia> adj_close = rand(5, 100);

julia> model = cornk(adj_close, 10, 3, 5, 3);

julia> model.type
"CORN-K"

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.cornu — Method

cornu(
  adj_close::Matrix{T},
  horizon::M,
  w::M;
  rho::T,
  init_budg=1
) where {T<:Float64, M<:Int}

Run CORN-U algorithm.

Arguments

adj_close::Matrix{T}: Adjusted close prices of assets.

horizon::M: The number of periods to invest.

w::M: maximum length of time window to be examined.

rho::T=0.2: The correlation coefficient threshold.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Reference

[1] CORN: Correlation-driven nonparametric learning approach for portfolio selection

Examples

julia> using OnlinePortfolioSelection

julia> adj_close = rand(5, 100);

julia> model = cornu(adj_close, 10, 5, 0.5);

julia> model.type
"CORN-U"

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.crp — Method

crp(adj_close::Matrix{T}, init_budg=1) where T<:Float64

Run Constant Rebalanced Portfolio (CRP) algorithm.

Arguments

adj_close::Matrix{Float64}: adjusted close prices

init_budg::Float64: initial budget

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): OPSAlgorithm object

References

[1] Universal Portfolios

Example

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> m_crp = crp(adj_close);

julia> m_crp.b
3×10 Matrix{Float64}:
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333

julia> sum(m_crp.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.dricornk — Method

dricornk(
  adj_close::Matrix{T},
  adj_close_market::Vector{T},
  horizon::M,
  k::M,
  w::M,
  p::M;
  lambda::T=1e-3,
  init_budg=1
) where {T<:Float64, M<:Int}

Run the DRICORNK algorithm.

Arguments

adj_close::Matrix{T}: A matrix of adjusted close prices of the assets.

adj_close_market::Vector{T}: A vector of adjusted close prices of the market in the same period.

horizon::M: The investment horizon.

k::M: The number of experts.

w::M: maximum length of time window to be examined.

p::M: maximum number of correlation coefficient thresholds.

lambda::T=1e-3: The regularization parameter.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Reference

[1] DRICORN-K: A Dynamic RIsk CORrelation-driven Non-parametric Algorithm for Online Portfolio Selection

Example

julia> using OnlinePortfolioSelection

julia> stocks_adj, market_adj = rand(10, 100), rand(100);

julia> m_dricornk = dricornk(stocks_adj, market_adj, 5, 2, 4, 3);

julia> sum(m_dricornk.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.eg — Method

eg(adj_close::Matrix{Float64}, init_budg=1., eta=0.05)

Exponential Gradient (EG) algorithm.

Calculate the Exponential Gradient (EG) weights and budgets using the given historical prices and parameters and return an EG object.

Arguments

adj_close::Matrix{Float64}: Historical adjusted close prices.

init_budg=1.: Initial budget.

eta=0.05: Learning rate.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): OPSAlgorithm object.

References

[1] On-Line Portfolio Selection Using Multiplicative Updates](https://onlinelibrary.wiley.com/doi/10.1111/1467-9965.00058)

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> m_eg = eg(adj_close);

julia> m_eg.b
3×10 Matrix{Float64}:
 0.333333  0.333119  0.333296  0.333232  0.33327   0.333276  0.333201  0.333171  0.332832  0.332789
 0.333333  0.333436  0.333274  0.333485  0.333481  0.333359  0.333564  0.333477  0.333669  0.333835
 0.333333  0.333445  0.33343   0.333283  0.333249  0.333365  0.333234  0.333353  0.333499  0.333377

julia> sum(m_eg.b, dims=1) .|> isapprox(1.0) |> all
true

OnlinePortfolioSelection.mdd — Method

mdd(Sn::Vector{T}) where T<:Float64

Calculate the Maximum Drawdown (MDD) of investment.

Arguments

Sn::Vector{T}: the cumulative return of investment during the investment period.

Returns

::Float64: the MDD of investment.

OnlinePortfolioSelection.sn — Method

sn(weights::Matrix{T}, rel_pr::Matrix{T}; init_inv::T=1.) where T<:Float64

Calculate the cumulative return of investment during a period of time.

Arguments

weights::Matrix{T}: the weights of the portfolio.

rel_pr::Matrix{T}: the relative price of the stocks.

init_inv::T=1: the initial investment.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

all_sn::Vector{T}: the cumulative return of investment during the investment period.

OnlinePortfolioSelection.up — Method

up(
  adj_close::Matrix{Float64};
  init_budg=1.,
  eval_points::Int=10^4,
  leverage=1.,
  frequency::Int=1,
  min_history::Int=0
)

Universal Portfolio (UP) algorithm.

Calculate the Universal Portfolio (UP) weights and budgets using the given historical prices and parameters.

Arguments

adj_close::Matrix{Float64}: Historical adjusted close prices.

init_budg=1.: Initial budget.

eval_points::Int=10^4: Number of evaluation points.

leverage=1.: Leverage value.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): OPSAlgorithm object.

References

[1] Universal Portfolios

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 30))

julia> m_up = up(adj_close);

julia> m_up.b
3×30 Matrix{Float64}:
 0.333333  0.331149  0.33204   0.331716  …  0.326788  0.325788  0.325829  0.326222
 0.333333  0.336058  0.335239  0.336304     0.343405  0.342161  0.342283  0.340693
 0.333333  0.332793  0.33272   0.331981     0.329807  0.332051  0.331888  0.333086

julia> sum(m_up.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.OPSAlgorithm — Type

OPSAlgorithm{T<:Float64}

A OPSAlgorithm object that contains the result of running the algorithm.

Fields

n_asset::Int: Number of assets in the portfolio.

b::Matrix{T}: Weights of the created portfolios.

alg::String: Name of the algorithm.

The formula for calculating the cumulative return of the portfolio is as follows:

\[{S_n} = {S_0}\prod\limits_{t = 1}^T {\left\langle {{b_t},{x_t}} \right\rangle }\]

where $S₀$ is the initial budget, $n$ is the investment horizon, $b_t$ is the vector of weights of the period $t$, and $x_t$ is the relative price of the $t$-th period.

OnlinePortfolioSelection.OPSMetrics — Type

OPSMetrics(Sn::Vector{T}, APY::T, Ann_Sharpe::T, MDD::T, Calmar::T) where {T<:Float64}

A struct to store the metrics of the OPS algorithm.

Fields

Sn::Vector{T}: the cumulative return of investment during the investment period.

APY::T: the Annual Percentage Yield (APY) of investment.

Ann_Std::T: the Annualized Standard Deviation (σₚ) of investment.

Ann_Sharpe::T: the Annualized Sharpe Ratio (SR) of investment.

MDD::T: the Maximum Drawdown (MDD) of investment.

Calmar::T: the Calmar Ratio of investment.

OnlinePortfolioSelection.OPSMetrics — Method

OPSMetrics(
  weights::Matrix{T},
  rel_pr::Matrix{T};
  init_inv::T=1.,
  Rf::T=0.02
  dpy::S=252
) where {T<:Float64, S<:Int}

Calculate the metrics of an OPS algorithm.

Arguments

weights::Matrix{T}: the weights of the portfolio.

rel_pr::Matrix{T}: the relative price of the stocks.

init_inv::T=1: the initial investment.

Rf::T=0.02: the risk-free rate of return.

dpy::S=252: the number of days in a year.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

::OPSMetrics: the metrics of the OPS algorithm.