API

OnlinePortfolioSelection.ann_sharpe — Method

ann_sharpe(APY::T, Rf::T, sigma_prtf::T) where T<:AbstractFloat

Calculate the Annualized Sharpe Ratio of investment.

Arguments

APY::T: the APY of investment.

Rf::T: the risk-free rate of return.

sigma_prtf::T: the standard deviation of the portfolio $sigma_p$.

Returns

::AbstractFloat: the Annualized Sharpe Ratio of investment.

OnlinePortfolioSelection.ann_std — Method

ann_std(cum_ret::Vector{AbstractFloat}; dpy)

Calculate the Annualized Standard Deviation (σₚ) of portfolio.

Arguments

cum_ret::Vector{AbstractFloat}: the cumulative return of investment during the investment period.

Keyword Arguments

dpy: the number of days in a year.

Returns

::AbstractFloat: the Annualized Standard Deviation (σₚ) of portfolio.

OnlinePortfolioSelection.anticor — Method

anticor(adj_close::Matrix{T}, window::Int) where {T<:Real}

Run the Anticor algorithm on adj_close with window sizes window.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Arguments

adj_close::Matrix{T}: matrix of adjusted close prices

window::Int: size of the window

Returns

::OPSAlgorithm(n_assets, b, alg): An OPSAlgorithm object.

Example

julia> using OnlinePortfolioSelection

julia> adj_close = [
       1. 2.
       4. 9.
       7. 8.
       10. 11.
       13. 7.
       8. 17.
       19. 20.
       22. 23.
       25. 8.
       2. 12.
       5. 12.
       5. 0.
       0. 2.
       1. 1.
       ];

julia> adj_close = permutedims(adj_close);

julia> m_anticor = anticor(adj_close, 3);

julia> m_anticor.b
2×14 Matrix{Float64}:
 0.5  0.5  0.5  0.5  …  0.0  0.0  0.0  1.0
 0.5  0.5  0.5  0.5     1.0  1.0  1.0  0.0

julia> sum(m_anticor.b, dims=1) .|> isapprox(1., atol=1e-8) |> all
true

References

[1] Can We Learn to Beat the Best Stock

OnlinePortfolioSelection.apy — Method

apy(Sn::AbstractFloat, n_periods::S; dpy::S=252) where S<:Int

Calculate the Annual Percentage Yield (APY) of investment.

Arguments

Sn::AbstractFloat: the cumulative return of investment.

n_periods::S: the number investment periods.

dpy::S=252: the number of days in a year.

Returns

::AbstractFloat: the APY of investment.

OnlinePortfolioSelection.bk — Method

bk(rel_price::Matrix{T}, K::S, L::S, c) where {T<:Float64, S<:Int}

Run Bᴷ algorithm.

Arguments

rel_price::Matrix{T}: Relative prices of assets.

K::S: Number of experts.

L::S: Number of time windows.

c::T: The similarity threshold.

Beware!

rel_price should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Example

julia> using OnlinePortfolioSelection

julia> daily_relative_prices = rand(10, 100);

julia> model = bk(daily_relative_prices, 10, 10, 0.5);

julia> model.alg
"Bᵏ"

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

Reference

[1] NONPARAMETRIC KERNEL-BASED SEQUENTIAL INVESTMENT STRATEGIES

OnlinePortfolioSelection.bs — Method

bs(adj_close::Matrix{T}; last_n::Int=0) where {T<:Float64}

Run the Best So Far algorithm on the given data.

Arguments

adj_close::Matrix{T}: A matrix of adjusted closing prices of assets.

Keyword Arguments

last_n::Int: The number of periods to look back for the performance of each asset. If last_n is 0, then the performance is calculated from the first period to the previous period.

Beware!

The adj_close matrix should be in the order of assets x periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An instance of OPSAlgorithm.

Example

julia> using OnlinePortfolioSelection

julia> adj_close = rand(5, 10);

julia> model = bs(adj_close, last_n=2);

julia> model.b
5×10 Matrix{Float64}:
 0.2  0.0  0.0  0.0  1.0  0.0  0.0  0.0  0.0  0.0
 0.2  0.0  0.0  1.0  0.0  1.0  0.0  0.0  0.0  0.0
 0.2  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0  0.0
 0.2  1.0  1.0  0.0  0.0  0.0  1.0  0.0  1.0  1.0
 0.2  0.0  0.0  0.0  0.0  0.0  0.0  1.0  0.0  0.0

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

References

[1] KERNEL-BASED SEMI-LOG-OPTIMAL EMPIRICAL PORTFOLIO SELECTION STRATEGIES

OnlinePortfolioSelection.calmar — Method

calmar(APY::T, MDD::T) where T<:AbstractFloat

Calculate the Calmar Ratio of investment.

Arguments

APY::T: the APY of investment.

MDD::T: the MDD of investment.

Returns

::AbstractFloat: the Calmar Ratio of investment.

OnlinePortfolioSelection.cluslog — Function

cluslog()
cluslog(
  rel_pr::AbstractMatrix{<:AbstractFloat},
  horizon::Int,
  TW::Int,
  clus_mod::Type{<:ClusteringModel},
  nclusters::Int,
  nclustering::Int,
  boundries::NTuple{2, AbstractFloat};
  log::Bool=true
)

You need to install and import the following packages before using this function:

Clustering

julia> using Pkg

julia> Pkg.add(name="Clustering", version="0.15.2")

julia> using OnlinePortfolioSelection, Clustering

cluslog(
  rel_pr::AbstractMatrix{<:AbstractFloat},
  horizon::Int,
  TW::Int,
  clus_mod::Type{<:ClusteringModel},
  nclusters::Int,
  nclustering::Int,
  boundries::NTuple{2, AbstractFloat};
  log::Bool=true
)

Run KMNLOG, KMDLOG, etc., algorithms on the given data.

Arguments

rel_pr::AbstractMatrix{<:AbstractFloat}: Relative prices of assets. Each column

represents the price of an asset at a given time.

horizon::Int: Number of trading days.

TW::Int: Maximum time window length to be examined.

clus_mod::Type{<:ClusteringModel}: Clustering model to be used. Currently, only

KmeansModel and KmedoidsModel are supported.

nclusters::Int: The maximum number of clusters to be examined.

nclustering::Int: The number of times clustering algorithm is run for optimal

number of clusters.

boundries::NTuple{2, AbstractFloat}: The lower and upper boundries for the

weights of assets in the portfolio.

Keyword Arguments

log::Bool=true: Whether to log the progress or not.

Beware!

rel_pr should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm: An OPSAlgorithm object.

Example

Two clustering model is available as of now: KmeansModel, and KmedoidsModel. The first example utilizes KmeansModel:

julia> using OnlinePortfolioSelection, Clustering

julia> adj_close = [
         1.5464 1.5852 1.6532 1.7245 1.5251 1.4185 1.2156 1.3231 1.3585 1.4563 1.4456
         1.2411 1.2854 1.3456 1.4123 1.5212 1.5015 1.4913 1.5212 1.5015 1.4913 1.5015
         1.3212 1.3315 1.3213 1.3153 1.3031 1.2913 1.2950 1.2953 1.3315 1.3213 1.3315
       ]

julia> rel_pr = adj_close[:, 2:end]./adj_close[:, 1:end-1]

julia> horizon = 3; TW = 3; nclusters_ = 3; nclustering = 10; lb, ub = 0.0, 1.;

julia> model = cluslog(rel_pr, horizon, TW, KmeansModel, nclusters_, nclustering, (lb, ub));

julia> model.b
3×3 Matrix{Float64}:
0.00264911  0.00317815  0.148012
0.973581    0.971728    0.848037
0.02377     0.0250939   0.00395028

julia> sum(model.b , dims=1) .|> isapprox(1.) |> all
true

The same approach works for KmedoidsModel as well:

julia> using OnlinePortfolioSelection, Clustering

julia> model = cluslog(rel_pr, horizon, TW, KmedoidsModel, nclusters_, nclustering, (lb, ub));

julia> model.b
3×3 Matrix{Float64}:
4.59938e-7  4.96421e-7  4.89426e-7
0.999998    0.999997    0.999997
2.02964e-6  2.02787e-6  2.02964e-6

julia> sum(model.b , dims=1) .|> isapprox(1.) |> all
true

Reference

An online portfolio selection algorithm using clustering approaches and considering transaction costs

OnlinePortfolioSelection.cornk — Method

cornk(
  adj_close::Matrix{Float64},
  horizon::T,
  k::T,
  w::T,
  p::T;
  init_budg=1
) where T<:Int

Run CORN-K algorithm.

Arguments

adj_close::Matrix{Float64}: Adjusted close prices of assets.

horizon::T: The number of periods to invest.

k::T: The number of top experts to be selected.

w::T: maximum length of time window to be examined.

p::T: maximum number of correlation coefficient thresholds.

Keyword Arguments

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Reference

[1] CORN: Correlation-driven nonparametric learning approach for portfolio selection

Examples

julia> using OnlinePortfolioSelection

julia> adj_close = rand(5, 100);

julia> model = cornk(adj_close, 10, 3, 5, 3);

julia> model.alg
"CORN-K"

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.cornu — Method

cornu(
  adj_close::Matrix{T},
  horizon::M,
  w::M;
  rho::T=0.2,
  init_budg=1
) where {T<:Float64, M<:Int}

Run CORN-U algorithm.

Arguments

adj_close::Matrix{T}: Adjusted close prices of assets.

horizon::M: The number of periods to invest.

w::M: maximum length of time window to be examined.

Keyword Arguments

rho::T=0.2: The correlation coefficient threshold.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Examples

julia> using OnlinePortfolioSelection

julia> adj_close = rand(5, 100);

julia> model = cornu(adj_close, 10, 5, 0.5);

julia> model.alg
"CORN-U"

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

Reference

[1] CORN: Correlation-driven nonparametric learning approach for portfolio selection

OnlinePortfolioSelection.crp — Method

crp(adj_close::Matrix{T}) where T<:Float64

Run Constant Rebalanced Portfolio (CRP) algorithm.

Arguments

adj_close::Matrix{Float64}: adjusted close prices

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): OPSAlgorithm object

Example

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> m_crp = crp(adj_close);

julia> m_crp.b
3×10 Matrix{Float64}:
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333
 0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333  0.333333

julia> sum(m_crp.b, dims=1) .|> isapprox(1.) |> all
true

References

[1] Universal Portfolios

OnlinePortfolioSelection.cwogd — Method

cwogd(
  rel_pr::AbstractMatrix,
  γ::AbstractFloat,
  H;
  bj::AbstractMatrix=diagm(ones(size(rel_pr, 1)))
)

Run the CW-OGD algorithm.

Positional Arguments

rel_pr::AbstractMatrix: Relative price matrix where it represents proportion of the closing price to the opening price of each asset in each day.

γ::AbstractFloat: Regular term coefficient of the basic expert's loss function.

H::AbstractFloat: Constant for calculating step sizes.

Keyword Arguments

bj::AbstractMatrix=diagm(ones(size(rel_pr, 1))): Matrix of experts opinions. Each column of this matrix must have just one positive element == 1. and others are zero. Also, sum of each column must be equal to 1. and number of rows must be equal to number of rows of rel_pr.

Beware!

rel_pr should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm: An object of OPSAlgorithm type.

Example

julia> using OnlinePortfolioSelection, YFinance

julia> tickers = ["AAPL", "MSFT", "GOOG"];

julia> startdt, enddt = "2019-01-01", "2019-01-10";

julia> querry_open_price = [get_prices(ticker, startdt=startdt, enddt=enddt)["open"] for ticker in tickers];

julia> open_pr = reduce(hcat, querry_open_price) |> permutedims;

julia> querry_close_pr = [get_prices(ticker, startdt=startdt, enddt=enddt)["adjclose"] for ticker in tickers];

julia> close_pr = reduce(hcat, querry_close_pr) |> permutedims;

julia> rel_pr = close_pr ./ open_pr
3×6 Matrix{Float64}:
 1.01956  0.987568  1.02581  0.994822  1.00796   1.01335
 1.01577  0.973027  1.02216  1.00413   0.997671  1.00395
 1.0288   0.976042  1.03692  0.997097  1.00016   0.993538

julia> gamma = 0.1; H = 0.5;

julia> model = cwogd(rel_pr, gamma, H);

julia> model.b
3×5 Matrix{Float64}:
 0.333333  0.351048  0.346241  0.338507  0.350524
 0.333333  0.321382  0.309454  0.320351  0.311853
 0.333333  0.32757   0.344305  0.341142  0.337623

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

Or using a custom matrix of experts opinions:

julia> b1 = [
          0.0 1.0 0.0
          1.0 0.0 0.0
          0.0 0.0 1.0
        ]

julia> model = cwogd(rel_pr, gamma, H, bj=b1);

julia> model.b
3×6 Matrix{Float64}:
 0.333333  0.329802  0.347517  0.34271   0.334976  0.346992
 0.333333  0.322351  0.3104    0.298472  0.309369  0.300871
 0.333333  0.347847  0.342083  0.358819  0.355655  0.352137

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

References

[1] Combining expert weights for online portfolio selection based on the gradient descent algorithm.

OnlinePortfolioSelection.dricornk — Method

dricornk(
  adj_close::Matrix{T},
  adj_close_market::Vector{T},
  horizon::M,
  k::M,
  w::M,
  p::M;
  lambda::T=1e-3,
  init_budg=1
) where {T<:Float64, M<:Int}

Run the DRICORNK algorithm.

Arguments

adj_close::Matrix{T}: A matrix of adjusted close prices of the assets.

adj_close_market::Vector{T}: A vector of adjusted close prices of the market in the same period.

horizon::M: The investment horizon.

k::M: The number of experts.

w::M: maximum length of time window to be examined.

p::M: maximum number of correlation coefficient thresholds.

Keyword Arguments

lambda::T=1e-3: The regularization parameter.

init_budg=1: The initial budget for investment.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Example

julia> using OnlinePortfolioSelection

julia> stocks_adj, market_adj = rand(10, 100), rand(100);

julia> m_dricornk = dricornk(stocks_adj, market_adj, 5, 2, 4, 3);

julia> sum(m_dricornk.b, dims=1) .|> isapprox(1.) |> all
true

Reference

[1] DRICORN-K: A Dynamic RIsk CORrelation-driven Non-parametric Algorithm for Online Portfolio Selection

OnlinePortfolioSelection.eg — Method

eg(adj_close::Matrix{Float64}; eta=0.05)

Exponential Gradient (EG) algorithm.

Calculate the Exponential Gradient (EG) weights and budgets using the given historical prices and parameters and return an EG object.

Arguments

adj_close::Matrix{Float64}: Historical adjusted close prices.

Keyword Arguments

eta=0.05: Learning rate.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): OPSAlgorithm object.

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 10))

julia> m_eg = eg(adj_close);

julia> m_eg.b
3×10 Matrix{Float64}:
 0.333333  0.333119  0.333296  0.333232  0.33327   0.333276  0.333201  0.333171  0.332832  0.332789
 0.333333  0.333436  0.333274  0.333485  0.333481  0.333359  0.333564  0.333477  0.333669  0.333835
 0.333333  0.333445  0.33343   0.333283  0.333249  0.333365  0.333234  0.333353  0.333499  0.333377

julia> sum(m_eg.b, dims=1) .|> isapprox(1.0) |> all
true

References

[1] On-Line Portfolio Selection Using Multiplicative Updates

OnlinePortfolioSelection.load — Method

load(adj_close::AbstractMatrix{T}, α::T, ω::S, horizon::S, η::T, ϵ::T=1.5) where {T<:Float64, S<:Int}

Run LOAD algorithm.

Arguments

adj_close::AbstractMatrix{T}: Adjusted close price data.

α::T: Decay factor. (0 < α < 1)

ω::S: Window size. (ω > 0)

horizon::S: Investment horizon. (n_periods > horizon > 0)

η::T: Threshold value. (η > 0)

ϵ::T=1.5: Expected return threshold value.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

OPSAlgorithm: An object of type OPSAlgorithm containing the weights of each asset for each period.

Sₜ::Vector{Float64}: Cumulative wealth for each period.

Example

# Get data
julia> using YFinance
julia> startdt, enddt = "2022-04-01", "2023-04-27";
julia> querry = [
          get_prices(ticker, startdt=startdt, enddt=enddt)["adjclose"] for ticker in tickers
       ];
julia> prices = reduce(hcat, querry);
julia> prices = permutedims(prices);

julia> using OnlinePortfolioSelection

julia> model, s = load(prices, 0.5, 30, 5, 0.1);

julia> model.b
5×5 Matrix{Float64}:
 0.2  2.85298e-8  0.0        0.0       0.0
 0.2  0.455053    0.637299   0.694061  0.653211
 0.2  0.215388    0.0581291  0.0       0.0
 0.2  0.329559    0.304572   0.305939  0.346789
 0.2  6.06128e-9  0.0        0.0       0.0

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

julia> s
6-element Vector{Float64}:
 1.0
 0.9879822754225864
 0.9853561439014098
 0.9836737048568326
 0.971437501096619
 0.9660091217094392

References

A local adaptive learning system for online portfolio selection

OnlinePortfolioSelection.mdd — Method

mdd(Sn::Vector{T}) where T<:AbstractFloat

Calculate the Maximum Drawdown (MDD) of investment.

Arguments

Sn::Vector{T}: the cumulative return of investment during the investment period.

Returns

::AbstractFloat: the MDD of investment.

OnlinePortfolioSelection.mer — Method

mer(
  weights::AbstractMatrix{T},
  rel_pr::AbstractMatrix{T},
  𝘷::AbstractFloat=0.
) where T<:AbstractFloat

Calculate the investments's Mean excess return (MER).

Arguments

weights::T: the weights of the portfolio.

rel_pr::T: the relative price of the stocks.

𝘷::AbstractFloat=0.: the transaction cost rate.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

MER::AbstractFloat: the investments's Mean excess return (MER).

OnlinePortfolioSelection.mrvol — Method

mrvol(
  rel_pr::AbstractMatrix{T},
  rel_vol::AbstractMatrix{T},
  horizon::S,
  Wₘᵢₙ::S,
  Wₘₐₓ::S,
  λ::T,
  η::T
) where {T<:AbstractFloat, S<:Integer}

Run MRvol algorithm.

Arguments

rel_pr::AbstractMatrix{T}: Relative price matrix where it represents proportion of the closing price to the opening price of each asset in each day.

rel_vol::AbstractMatrix{T}: Relative volume matrix where 𝘷ᵢⱼ represents the tᵗʰ trading volume of asset 𝑖 divided by the (t - 1)ᵗʰ trading volume of asset 𝑖.

horizon::S: Investment horizon. The last horizon days of the data will be used to run the algorithm.

Wₘᵢₙ::S: Minimum window size.

Wₘₐₓ::S: Maximum window size.

λ::T: Trade-off parameter in the loss function.

η::T: Learning rate.

Beware!

rel_pr and rel_vol should be matrixes of size n_assets × n_periods.

Returns

OPSAlgorithm: An OPSAlgorithm object.

Example

julia> using OnlinePortfolioSelection, YFinance

julia> tickers = ["AAPL", "MSFT", "GOOG"];

julia> startdt, enddt = "2019-01-01", "2020-01-01";

julia> querry_open_price = [get_prices(ticker, startdt=startdt, enddt=enddt)["open"] for ticker in tickers];

julia> open_pr = reduce(hcat, querry_open_price) |> permutedims;

julia> querry_close_pr = [get_prices(ticker, startdt=startdt, enddt=enddt)["adjclose"] for ticker in tickers];

julia> close_pr = reduce(hcat, querry_close_pr) |> permutedims;

julia> querry_vol = [get_prices(ticker, startdt=startdt, enddt=enddt)["vol"] for ticker in tickers];

julia> vol = reduce(hcat, querry_vol) |> permutedims;

julia> rel_pr = (close_pr ./ open_pr)[:, 2:end];

julia> rel_vol = vol[:, 2:end] ./ vol[:, 1:end-1];

julia> size(rel_pr) == size(rel_vol)
true

julia> horizon = 100; Wₘᵢₙ = 4; Wₘₐₓ = 10; λ = 0.05; η = 0.01;

julia> r = mrvol(rel_pr, rel_vol, horizon, Wₘᵢₙ, Wₘₐₓ, λ, η);

julia> r.b
3×100 Matrix{Float64}:
 0.333333  0.0204062  0.0444759  …  0.38213   0.467793
 0.333333  0.359864   0.194139      0.213264  0.281519
 0.333333  0.61973    0.761385      0.404606  0.250689

References

[1] Online portfolio selection of integrating expert strategies based on mean reversion and trading volume.

OnlinePortfolioSelection.olmar — Method

olmar(adj_close::Matrix{Float64}, ϵ::Int, ω::Int)::OPSAlgortihm

Run the Online Moving Average Reversion algorithm.

Arguments

adj_close::Matrix{Float64}: matrix of adjusted closing prices.

ϵ::Int: Reversion threshold.

ω::Int: Window size.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgortihm(n_asset, b, alg): An OPSAlgortihm object.

Example

julia> using OnlinePortfolioSelection

julia> adj = [
        1.315 1.326 1.358 1.39461 1.424 1.4015 1.52531
        1.215 1.111 1.215 1.35614 1.454 1.2158 1.32561
       ];

julia> m_olmar = olmar(adj, 2, 3);

julia> m_olmar.b
2×7 Matrix{Float64}:
 0.5  0.5  1.0  1.0  1.0  1.0  0.0
 0.5  0.5  0.0  0.0  0.0  0.0  1.0

julia> all(sum(m_olmar.b, dims=1) .≈ 1.0)
true

References

[1] On-Line Portfolio Selection with Moving Average Reversion

OnlinePortfolioSelection.opsmethods — Method

opsmethods()

Print the available methods in the package.

Example

julia> using OnlinePortfolioSelection

julia> opsmethods()

      ===== OnlinePortfolioSelection.jl =====
            Currently available methods
       =====================================

        up: Universal Portfolio - Call `up`
        eg: Exponential Gradient - Call `eg`
     cornu: CORN-U - Call `cornu`
          ⋮

OnlinePortfolioSelection.rprt — Method

rprt(
  adj_close::Matrix{T};
  w::Int64=5,
  theta::T=0.8,
  epsilon=50
) where T<:Float64

Run RPRT algorithm.

Arguments

adj_close::Matrix{T}: Adjusted close prices of assets.

Keyword Arguments

w::Int64=5: maximum length of time window to be examined.

theta::T=0.8: The threshold for the relative price.

epsilon=50: The threshold for the condition of the portfolio.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): An object of type OPSAlgorithm.

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 7))

julia> m_rprt = rprt(adj_close);

julia> m_rprt.b
3×7 Matrix{Float64}:
 0.333333  0.333333  0.0  0.0  0.0  1.0  0.0
 0.333333  0.333333  1.0  1.0  1.0  0.0  1.0
 0.333333  0.333333  0.0  0.0  0.0  0.0  0.0

julia> sum(m_rprt.b, dims=1) .|> isapprox(1.) |> all
true

Reference

[1] Reweighted Price Relative Tracking System for Automatic Portfolio Optimization

OnlinePortfolioSelection.sn — Method

sn(weights::AbstractMatrix{T}, rel_pr::AbstractMatrix{T}; init_inv::T=1.) where T<:AbstractFloat

Calculate the cumulative return of investment during a period of time.

Arguments

weights::AbstractMatrix{T}: the weights of the portfolio.

rel_pr::AbstractMatrix{T}: the relative price of the stocks.

Keyword Arguments

init_inv::T=1: the initial investment.

Beware!

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

all_sn::Vector{T}: the cumulative return of investment during the investment period.

OnlinePortfolioSelection.uniform — Method

uniform(n_assets::Int, horizon::Int)

Construct uniform portfolios.

Arguments

n_assets::Int: The number of assets.

horizon::Int: The number of investment periods.

Returns

::OPSAlgorithm: An object of OPSAlgorithm type.

Example

julia> using OnlinePortfolioSelection

julia> model = uniform(3, 10)

julia> sum(model.b, dims=1) .|> isapprox(1.) |> all
true

OnlinePortfolioSelection.up — Method

up(adj_close::Matrix{Float64}; eval_points::Int=10^4)

Universal Portfolio (UP) algorithm.

Calculate the Universal Portfolio (UP) weights and budgets using the given historical prices and parameters.

Arguments

adj_close::Matrix{Float64}: Historical adjusted close prices.

Keyword Arguments

eval_points::Int=10^4: Number of evaluation points.

Beware!

adj_close should be a matrix of size n_assets × n_periods.

Returns

::OPSAlgorithm(n_assets, b, alg): OPSAlgorithm object.

Examples

julia> using OnlinePortfolioSelection

julia> typeof(adj_close), size(adj_close)
(Matrix{Float64}, (3, 30))

julia> m_up = up(adj_close);

julia> m_up.b
3×30 Matrix{Float64}:
 0.333333  0.331149  0.33204   0.331716  …  0.326788  0.325788  0.325829  0.326222
 0.333333  0.336058  0.335239  0.336304     0.343405  0.342161  0.342283  0.340693
 0.333333  0.332793  0.33272   0.331981     0.329807  0.332051  0.331888  0.333086

julia> sum(m_up.b, dims=1) .|> isapprox(1.) |> all
true

References

[1] Universal Portfolios

OnlinePortfolioSelection.OPSAlgorithm — Type

OPSAlgorithm{T<:Float64}

A OPSAlgorithm object that contains the result of running the algorithm.

Fields

n_asset::Int: Number of assets in the portfolio.

b::Matrix{T}: Weights of the created portfolios.

alg::String: Name of the algorithm.

The formula for calculating the cumulative return of the portfolio is as follows:

\[{S_n} = {S_0}\prod\limits_{t = 1}^T {\left\langle {{b_t},{x_t}} \right\rangle }\]

where $S₀$ is the initial budget, $n$ is the investment horizon, $b_t$ is the vector of weights of the period $t$, and $x_t$ is the relative price of the $t$-th period.

OnlinePortfolioSelection.OPSMetrics — Type

OPSMetrics(Sn::Vector{T}, APY::T, Ann_Sharpe::T, MDD::T, Calmar::T) where {T<:AbstractFloat}

A struct to store the metrics of the OPS algorithm.

Fields

Sn::Vector{T}: the cumulative return of investment during the investment period.

APY::T: the Annual Percentage Yield (APY) of investment.

Ann_Std::T: the Annualized Standard Deviation (σₚ) of investment.

Ann_Sharpe::T: the Annualized Sharpe Ratio (SR) of investment.

MDD::T: the Maximum Drawdown (MDD) of investment.

Calmar::T: the Calmar Ratio of investment.

OnlinePortfolioSelection.OPSMetrics — Method

OPSMetrics(
  weights::AbstractMatrix{T},
  rel_pr::AbstractMatrix{T};
  init_inv::T=1.,
  Rf::T=0.02
  dpy::S=252,
  v::T=0.
  dpy::S=252
) where {T<:AbstractFloat, S<:Int}

Calculate the metrics of an OPS algorithm.

Arguments

weights::AbstractMatrix{T}: the weights of the portfolio.

rel_pr::AbstractMatrix{T}: the relative price of the stocks.

Keyword Arguments

init_inv::T=1: the initial investment.

Rf::T=0.02: the risk-free rate of return.

dpy::S=252: the number of days in a year.

v::T=0.: the transaction cost rate.

Warning

The size of weights and rel_pr must be (n_stocks, n_periods).

Note

If size(rel_pr, 2) is greater than size(weights, 2), then the last size(weights, 2) columns of rel_pr will be used.

Returns

::OPSMetrics: the metrics of the OPS algorithm.